New improvements of the mobile part. This includes: - Improved RESTFul JSON webservice (still, only mock-data) - Allows for transport of generic/unknown ELO content - A framework for serializing and deserializing JSON data on the mobile client
The TestELOServiceClient example MIDlet shows how to retrieve an ELO from the server, serialize it on the client side and display its content. Currently two different elo mockups are supported. One with plain text as its content and another with a list of images as its content. Test it by entering id=1 for the text-elo and id=3 for the image elo.